# Move to Iterative Trie implementation (`trieNode`)