--- a/lib/types/hostnametrie.go
+++ b/lib/types/hostnametrie.go
@@ -134,5 +134,6 @@ func (t *HostnameTrie) insert(s string) error {
 // Contains returns whether s matches a pattern in the HostnameTrie
 // along with the matching pattern, if one was found.
 func (t *HostnameTrie) Contains(s string) (matchedPattern string, matchFound bool) {
+	s = strings.ToLower(s)
 	return t.trieNode.contains(s)
 }