Hi @codebien 👋🏻 here is the aforementioned PR

What have changed:

- Moved `trieNode` to separate file
- New unit tests for trie implementation
- New benchmarks for the `trieNode.insert()` and `trieNode.contains()` methods
- Moved from recursive to iterative implementation

Here is a simple script that I used to create the benchmarks on my machine.

```bash
bench(){
  echo "started: $2 benchmark"
  git checkout "$1"  2> /dev/null
  go test -run=^$ -bench=BenchmarkTrie -benchmem -count 30 > "$2"
}

cd lib/types

bench 2f14099  "original"
bench e11e0679 "remove_error"
bench 08933201 "iterative"

benchstat original remove_error > "bs_remove_error"
benchstat original iterative > "bs_iterative"
```

# Benchmarks

My machine was idle.

```
goos: darwin
goarch: amd64
pkg: go.k6.io/k6/lib/types
cpu: Intel(R) Core(TM) i5-8257U CPU @ 1.40GHz
```

##  `bs_iterative`

```
name            old time/op    new time/op    delta
TrieInsert-8      13.3µs ± 1%     6.3µs ± 1%  -52.78%  (p=0.000 n=29+30)
TrieContains-8    11.7µs ± 2%     3.1µs ± 1%  -73.01%  (p=0.000 n=30+27)

name            old alloc/op   new alloc/op   delta
TrieInsert-8      7.46kB ± 0%    7.46kB ± 0%     ~     (all equal)
TrieContains-8    1.02kB ± 0%    0.42kB ± 0%  -58.27%  (p=0.000 n=30+30)

name            old allocs/op  new allocs/op  delta
TrieInsert-8         130 ± 0%       130 ± 0%     ~     (all equal)
TrieContains-8      90.0 ± 0%      24.0 ± 0%  -73.33%  (p=0.000 n=30+30)
```

##  `bs_remove_error`

```
name          old time/op    new time/op    delta
TrieInsert-8    14.0µs ± 0%    13.3µs ± 1%   ~     (p=0.106 n=1+28)

name          old alloc/op   new alloc/op   delta
TrieInsert-8    7.46kB ± 0%    7.46kB ± 0%   ~     (all equal)

name          old allocs/op  new allocs/op  delta
TrieInsert-8       130 ± 0%       130 ± 0%   ~     (all equal)
```