@eugercek 2022-11-18
IMO `trieNode` should be generic enough because it's a DS, it should be usable in other part of the codebase. Currently we only use in `HostnameTrie` and `AddresTrie` (not in master a separate PR). Because hostname is not case sensitive we lowercased it in `trieNode.contains()` too, last commit fixes that.