@eugercek 2022-11-21
Hi @imiric this PR was solely optional, I thought iterative one can be simpler and more performant. I think `insert` is simpler but I agree `contains` is more complex in iterative implementation. It's up to you guys.

Fixed the linter errors, sorry for that.

Also thanks for all of your kind words, it's my pleasure to ,trying to, contribute to k6.