@eugercek 2022-11-28
Only ` ST1000: at least one file in a package should have a package comment ` was producing error, added a little comment.